norm: fix norm
